Closer to finalizing on the bv_data_assim_errors.py. Recent changes:
1. Fixed in error in writeback to dS in the _I_ function and _J_ funciton. These
were wrong, but have been corrected.
2. Assimilation now done on ny rather than nx in the normalized direction
(nx,ny). This allows ny to change sign, which is important, given the
predominently EW direcion of Greenland ice flow.
3. Changed from ln to (Uobs - Ubmag)**2 in the objective function. I think that
the outlets *should* be more imporant than the interior.
4. Looking at about +- 15 degree errors in direction. High, but works...

Currently running on a very crude mesh and getting RMS errors of about 40 m/a.
Compare to 36 m/a reported in VarGlaS paper. Not bad!